gcp_product_curation/requirements.txt
# requirements.txt
httpx
beautifulsoup4
mcp
pg8000
google-cloud-alloydb-connector[pg8000]
pgvector
google-adk
fastapi
gunicorn
pydantic
psycopg2-binary
requests